+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * added the PDO Store
+ * Add a new Zookeeper Data Store for Lock Component.
 
 3.4.0
 -----